chore: bump version to v23.6.0 (Vega-W)
Release v23.6.0 Vega-W - Security Hardening v2 (Sprints 461-470)

- 8 new security headers: ZeroTrust, DecoderSandbox, RuntimeIntegrity, ExploitMitigation, PrivilegeSeparation, SecureIPC, AuditTrail, AntiTamper

- Test count: 3117 unit tests (+179)

- Added SPRINT_PLAN_700.md (Sprints 661-760, v26.2.0 Canopus-S through v27.3.0 Sirius-T)

- Updated all version references, README badge, SVG graphics, CHANGELOG

+++ Engine/Core/SecureStringPool.h
@@ -55,7 +55,7 @@ public:
         // Encrypt with CryptProtectMemory (CRYPTPROTECTMEMORY_SAME_PROCESS scope)
         // Block size for CryptProtectMemory must be multiple of CRYPTPROTECTMEMORY_BLOCK_SIZE (16)
         DWORD blockSize = 16;
-        size_t encLen = (byteLen + blockSize - 1) & ~(blockSize - 1);
+        size_t encLen = (byteLen + blockSize - 1) & ~(static_cast<size_t>(blockSize) - 1);
         if (encLen <= pageSize) {
             if (CryptProtectMemory(slot.page, static_cast<DWORD>(encLen),
                     CRYPT_PROTECT_MEMORY_SAME_PROCESS))
@@ -79,7 +79,7 @@ public:
         // Temporarily decrypt
         if (slot.encrypted) {
             DWORD blockSize = 16;
-            size_t encLen = (slot.dataLen + blockSize - 1) & ~(blockSize - 1);
+            size_t encLen = (slot.dataLen + blockSize - 1) & ~(static_cast<size_t>(blockSize) - 1);
             CryptUnprotectMemory(slot.page, static_cast<DWORD>(encLen),
                     CRYPT_PROTECT_MEMORY_SAME_PROCESS);
         }
@@ -89,7 +89,7 @@ public:
         // Re-encrypt immediately
         if (slot.encrypted) {
             DWORD blockSize = 16;
-            size_t encLen = (slot.dataLen + blockSize - 1) & ~(blockSize - 1);
+            size_t encLen = (slot.dataLen + blockSize - 1) & ~(static_cast<size_t>(blockSize) - 1);
             CryptProtectMemory(slot.page, static_cast<DWORD>(encLen),
                     CRYPT_PROTECT_MEMORY_SAME_PROCESS);
         }
